Added styling, added photo, fixed some bugs
Added multiple SASS files and implemented some styling for the page.
Fixed a bug where net-cashflow wasn't calculating properly.
Fixed a bug that gave the last recorded net-cashflow when a new user
signed in that didn't have any cash flows or when a user deleted the
last cashflow.
Fixed a bug in handlebars that improperly displayed a string, only showing
the first word.
Fixed a bug that would display net-cashflow even when there was nothing
to display.