Update works with table, added styling
Upon selecting a row the update form is automatically populated with
selected data. The user can then alter the data and click the update
button to update.
Added some images to the images directory.
Used one of the images for the body background.
Adjusted styling for increased visibility.
Improved the introduction paragraph.
Moved the update form html out of handlebars and into index.html.